+++ core-ng/src/main/java/core/framework/db/CloudAuthProvider.java
@@ -1,4 +1,4 @@
-package core.framework.internal.db.cloud;
+package core.framework.db;
 
 /**
  * @author neo
@@ -7,4 +7,8 @@ public interface CloudAuthProvider {
     String user();
 
     String accessToken();
+
+    class Registry {
+        public static CloudAuthProvider INSTANCE;
+    }
 }