+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 ## Change log
 
-### 9.0.2 (12/7/2023 - )
+### 9.0.2-b0 (12/7/2023 - )
 
 * stats: dump virtual threads on high cpu
 * http: response "connection: keep-alive" header if client sends keep-alive header
   > to be compatible with http/1.0 client, like ab (apache benchmark) with "-k"
+* mysql: patched mysql jdbc driver to support virtual thread and gcloud auth
+  > use "core.framework.mysql:mysql-connector-j:8.2.0-p1"
 
 ### 9.0.1 (12/01/2023 - 12/7/2023)